Refactor for class auto loader to work
