+++ backend/api/user/user_create_code.php
@@ -1,7 +1,6 @@
 <?php
 // Initiate a new user account creation
 include_once('../../lib/tools.php');
-include_once('../../lib/mail.php');
 $json = new JSON_Resp();
 
 define("DEBUG", False);
